test: consolidating test
